Add update-trilio action
Add update-trilio action which can be run after a Trilio
upgrade.

Trilio sometimes updates packages in package repositiories that
require DB migrations or config file changes. Since these package
updates are not applied by the charm as they would be by changing the
openstack-origin or triliovault-pkg-source the charm has no event to
react to to automatically perform the post package upgrade actions.
This change adds an action `update-trilio` which can
be run to apply the changes.

Change-Id: I52abfa8091673f20c8c5cc1fffd7ad5d17a60820